Refactor web view activity
