--- ui/src/main/java/com/midtrans/sdk/ui/views/webpayment/MidtransWebviewFragment.java
@@ -1,112 +1,0 @@
-package com.midtrans.sdk.ui.views.webpayment;
-
-import android.annotation.SuppressLint;
-import android.app.Activity;
-import android.content.Intent;
-import android.os.Build;
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.webkit.WebChromeClient;
-import android.webkit.WebSettings;
-import android.webkit.WebView;
-
-import com.midtrans.sdk.ui.R;
-import com.midtrans.sdk.ui.abtracts.BaseFragment;
-import com.midtrans.sdk.ui.utils.Logger;
-import com.midtrans.sdk.ui.utils.UiUtils;
-
-public class MidtransWebviewFragment extends BaseFragment implements MidtransWebViewClient.WebViewClientCallback {
-
-    private static final String URL_PARAM = "url_param";
-    private static final String TYPE_PARAM = "type_param";
-
-    public WebView webView;
-    private String webUrl;
-    private String type;
-
-    public MidtransWebviewFragment() {
-    }
-
-    public static MidtransWebviewFragment newInstance(String url) {
-        MidtransWebviewFragment fragment = new MidtransWebviewFragment();
-        Bundle args = new Bundle();
-        args.putString(URL_PARAM, url);
-
-        fragment.setArguments(args);
-        return fragment;
-    }
-
-    public static MidtransWebviewFragment newInstance(String url, String type) {
-        MidtransWebviewFragment fragment = new MidtransWebviewFragment();
-        Bundle args = new Bundle();
-        args.putString(URL_PARAM, url);
-        args.putString(TYPE_PARAM, type);
-        fragment.setArguments(args);
-        return fragment;
-    }
-
-    @Override
-    public void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        Logger.d(TAG, "onCreate");
-
-        if (getArguments() != null) {
-            webUrl = getArguments().getString(URL_PARAM);
-            type = getArguments().getString(TYPE_PARAM);
-            Logger.d(TAG, "weburl:" + webUrl);
-            Logger.d(TAG, "type:" + type);
-        }
-    }
-
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container,
-                             Bundle savedInstanceState) {
-        return inflater.inflate(R.layout.fragment_webview, container, false);
-    }
-
-    @Override
-    public void onViewCreated(View view, Bundle savedInstanceState) {
-        webView = (WebView) view.findViewById(R.id.webview);
-        initwebview();
-        webView.loadUrl(webUrl);
-        super.onViewCreated(view, savedInstanceState);
-    }
-
-    @SuppressLint("AddJavascriptInterface")
-    private void initwebview() {
-        UiUtils.showProgressDialog(getActivity(), true);
-        webView.getSettings().setAllowFileAccess(false);
-        webView.getSettings().setJavaScriptEnabled(true);
-        webView.setInitialScale(1);
-        webView.getSettings().setLoadWithOverviewMode(true);
-        webView.getSettings().setUseWideViewPort(true);
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-            webView.getSettings().setMixedContentMode(WebSettings.MIXED_CONTENT_ALWAYS_ALLOW);
-        }
-        webView.setWebViewClient(new MidtransWebViewClient(this, type));
-        webView.setWebChromeClient(new WebChromeClient());
-    }
-
-    @Override
-    public void onDetach() {
-        super.onDetach();
-        if (webView != null) {
-            webView.removeAllViews();
-            webView.destroyDrawingCache();
-            webView.clearHistory();
-            webView.destroy();
-        }
-    }
-
-    @Override
-    public void onCompleted() {
-        if (getActivity() != null) {
-            Intent returnIntent = new Intent();
-            getActivity().setResult(Activity.RESULT_OK, returnIntent);
-            getActivity().finish();
-        }
-    }
-
-}
