--- ui/src/main/java/com/midtrans/sdk/ui/views/webpayment/PaymentWebPresenter.java
@@ -1,10 +1,0 @@
-package com.midtrans.sdk.ui.views.webpayment;
-
-import com.midtrans.sdk.ui.abtracts.BasePaymentPresenter;
-
-/**
- * Created by rakawm on 4/6/17.
- */
-
-public class PaymentWebPresenter extends BasePaymentPresenter {
-}
