+++ ui/src/main/res/layout/activity_payment_web.xml
@@ -10,10 +10,11 @@
 
     <include layout="@layout/toolbar_payment_page" />
 
-    <RelativeLayout
-        android:id="@+id/webview_container"
+    <WebView
+        android:id="@+id/webview"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        app:layout_behavior="@string/appbar_scrolling_view_behavior" />
+        app:layout_behavior="@string/appbar_scrolling_view_behavior"
+        tools:context="com.midtrans.sdk.ui.views.webpayment.PaymentWebActivity" />
 
 </android.support.design.widget.CoordinatorLayout>
